fix(binary): Clarify length_data as length_take
`take` is generally the verb we use for creating slices, maing the
intent more obvious.
